requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
defusedxml==0.7.1
fonttools==4.47.2
fpdf2==2.7.7
idna==3.4
mss==9.0.1
olefile==0.47
pillow==10.2.0
printfactory==0.1.1
pydantic==1.10.14
PyPDF2==3.0.1
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.0.7
pywin32==306; platform_system=="Windows"
